refactor mobile notification view
