+++ web/src/components/NotificationDropdown/NotificaitonComponent.tsx
@@ -1,0 +1,91 @@
+import RoundAvatar from "../RoundAvatar/RoundAvatar";
+import ClickableId from "../ClickableId/ClickableId";
+import timePosted from "../../timePosted";
+import summary from "../../summary";
+import React from "react";
+import Notification from "../../models/Notification";
+import {useHistory} from "react-router-dom";
+
+interface Props {
+  notification: Notification
+  api: any
+}
+
+export default (props: Props) => {
+  let notifier
+  const notification = props.notification
+  const history = useHistory()
+
+  const notificationSummary = (notification: Notification) => {
+    const notification_summary = notification.notifying_summary
+    return summary(notification_summary, 100)
+  }
+
+  if (notification.notifying_action !== "mention") {
+    notifier = !notification.notifying_deleted ? notification.notifier : null
+  } else {
+    notifier = !notification.notified_deleted ? notification.notifier : null
+  }
+
+  let action
+  if (notification.notifying_action === "mention") action = "mentioned"
+  if (notification.notifying_action === "reshare") action = "reshared"
+  if (notification.notifying_action === "comment") action = "commented"
+  if (notification.notifying_action === "reaction") action =  "reacted"
+
+  let notifiedLocationPronoun
+  if (notification.notifying_action === "mention") notifiedLocationPronoun = "their"
+  else notifiedLocationPronoun = "your"
+
+  let notifiedLocationType
+  if (notification.notified_href.indexOf("#comment-") !== -1) notifiedLocationType = 'comment'
+  else if (notification.notified_href.indexOf("/post/") !== -1) notifiedLocationType = 'post'
+
+  const notificationOnClick = async () => {
+    await props.api.markNotificationAsRead(notification.id)
+    history.push(notification.notified_href)
+  }
+
+  return (
+    <div className="notification-wrapper" style={{
+      backgroundColor: notification.unread ? 'white' : '#f0f0f0'
+    }}>
+      <div className="notification-first-row" onClick={notificationOnClick}>
+        <div className="notification-info">
+          <div className="post-avatar notification-avatar">
+            <RoundAvatar user={notification.notifier}/>
+          </div>
+          <div className="notification-notifier">
+            <div className="notification-notifier-wrapper">
+              <b className="notification-notifier-id">
+                <ClickableId user={notifier}/>{' '}
+              </b>
+              {action}
+              {' '}
+              {
+                notification.notifying_action === "mention" ?
+                  "you" :
+                  <div className="notification-summary">
+                    "{notificationSummary(notification)}"
+                  </div>
+              }
+              {' '}
+              on
+              {' '}
+              {notifiedLocationPronoun}
+              {' '}
+              {notifiedLocationType}
+            </div>
+          </div>
+        </div>
+        <div className="notification-time">
+          {timePosted(notification.created_at_seconds)}
+        </div>
+      </div>
+
+      <div className="notification-second-row" onClick={notificationOnClick}>
+        {summary(notification.notified_summary, 150)}
+      </div>
+    </div>
+  )
+}
